Only execute scratchpad related logic, when scratchpad is visible.
